backend/requirement.txt
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataloader==2.0
filelock==3.13.1
Flask==3.0.3
fonttools==4.47.0
fsspec==2023.12.2
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
seaborn==0.13.0
six==1.16.0
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.3
